# specs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Teams API
  description: API for managing soccer teams, players, trainers, events, and equipment
  version: 1.0.0
  contact:
    name: Teams API Support
    email: support@teams-api.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.teams.com
    description: Production server
# Authentication is optional for most endpoints, but required for some
security: []
paths:
...
  /api/groups/{groupId}/members:
    get:
      operationId: getAllMembers
      summary: Get all members in a group
      description: Retrieve all members for a specific group in a single response (players and trainers). Guardians can read this endpoint, but sensitive player fields are hidden.
      tags:
        - Members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved members
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items:
                      $ref: '#/components/schemas/Player'
                  trainers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trainer'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createMember
      summary: Create a new member in a group
      description: Create either a player or trainer based on the role field in a specific group
      tags:
        - Members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreatePlayerRequest'
                - $ref: '#/components/schemas/CreateTrainerRequest'
      responses:
        '201':
          description: Member created successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PlayerWithRole'
                  - $ref: '#/components/schemas/TrainerWithRole'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/groups/{groupId}/members/{id}:
    get:
      operationId: getMemberById
      summary: Get member by ID
      description: Retrieve a specific member (player or trainer) by ID in a specific group. Guardians can read this endpoint, but sensitive player fields are hidden.
      tags:
        - Members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PlayerWithRole'
                  - $ref: '#/components/schemas/TrainerWithRole'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateMemberById
      summary: Update member by ID
      description: Update a specific Member (player or trainer) by ID in a specific group
      tags:
        - Members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/UpdatePlayerRequest'
                - $ref: '#/components/schemas/UpdateTrainerRequest'
      responses:
        '200':
          description: Member updated successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PlayerWithRole'
                  - $ref: '#/components/schemas/TrainerWithRole'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteMemberById
      summary: Delete member by ID
      description: Delete a specific Member (player or trainer) by ID in a specific group
      tags:
        - Members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Member deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to access this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Player:
      type: object
      properties:
        id:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/GroupRole'
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
          description: Canonical player birth date; birthYear is derived from this value.
        birthYear:
          type: integer
          readOnly: true
          deprecated: true
          description: Derived from birthDate. Use birthDate for writes.
        level:
          type: integer
          minimum: 1
          maximum: 5
        preferredShirtNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [active, trial, inactive]
          default: active
        guardians:
          type: array
          items:
            $ref: '#/components/schemas/Guardian'
      required:
        - id
        - roles
        - firstName
        - lastName
        - level
...
    Trainer:
      type: object
      properties:
        id:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/GroupRole'
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
          nullable: true
      required:
        - id
        - roles
        - firstName
        - lastName
    Guardian:
      type: object
      properties:
        id:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/GroupRole'
        groupId:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        userId:
          type: string
          deprecated: true
          description: Legacy guardian account identifier; use the guardian member id.
      required:
        - id
        - groupId
        - roles
    GroupRole:
      type: string
      enum: [admin, trainer, guardian, player]
...
    PlayerWithRole:
      allOf:
        - $ref: '#/components/schemas/Player'
    TrainerWithRole:
      allOf:
        - $ref: '#/components/schemas/Trainer'
    CreatePlayerRequest:
      type: object
      properties:
        role:
          type: string
          enum: [player]
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
          description: Canonical date of birth. Do not send the derived birthYear field.
        level:
          type: integer
          minimum: 1
          maximum: 5
        preferredShirtNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [active, trial, inactive]
          default: active
      required:
        - role
        - firstName
        - lastName
        - birthDate
        - level
    CreateTrainerRequest:
      type: object
      properties:
        role:
          type: string
          enum: [trainer]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
      required:
        - role
        - firstName
        - lastName
    UpdatePlayerRequest:
      type: object
      properties:
        role:
          type: string
          enum: [player]
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
          description: Canonical date of birth. Do not send the derived birthYear field.
        level:
          type: integer
          minimum: 1
          maximum: 5
        preferredShirtNumber:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [active, trial, inactive]
          default: active
      required:
        - role
        - firstName
        - lastName
        - birthDate
        - level
    UpdateTrainerRequest:
      type: object
      properties:
        role:
          type: string
          enum: [trainer]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
      required:
        - role
        - firstName
        - lastName
...